requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
python-decouple==3.6
pytz==2022.7
sqlparse==0.4.3
